refactor: action 감싸기
