--- mongeul/src/actions/diary/createDiaryDraft.ts
@@ -1,36 +1,0 @@
-"use server";
-
-import { revalidatePath } from "next/cache";
-import { DiaryRequest, DiaryResponse } from "@/types/diaryTypes";
-
-export async function createDiaryDraft(
-  data: DiaryRequest
-): Promise<DiaryResponse> {
-  try {
-    console.log("request data:", { data });
-    const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-
-    const response = await fetch(`${API_BASE_URL}/api/v1/diaries/drafts`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${process.env.NEXT_PUBLIC_API_TOKEN}`,
-      },
-      body: JSON.stringify(data),
-      cache: "no-store",
-    });
-
-    if (!response.ok) {
-      const errorMessage = await response.text();
-      throw new Error(`일기 임시저장 실패: ${errorMessage}`);
-    }
-
-    const result: DiaryResponse = await response.json();
-    // revalidatePath("/diary"); // 일기 목록 데이터 새로고침
-
-    return result;
-  } catch (error) {
-    console.error("일기 임시저장 에러 발생:", error);
-    throw error;
-  }
-}
